+++ api/mapper/oracle.go
@@ -5,7 +5,7 @@ import (
 	"math/big"
 	"strings"
 
-	"github.com/figment-networks/indexer-manager/structs"
+	"github.com/figment-networks/indexing-engine/structs"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/tendermint/tendermint/libs/bech32"