+++ api/types/typesV4.go
@@ -1,29 +1,50 @@
 package types
 
-// GetBlockchainResponseV4 cosmos response from blockchain
-type GetBlockchainResponseV4 struct {
-	ID     int                `json:"id"`
-	RPC    string             `json:"jsonrpc"`
-	Result ResultBlockchainV4 `json:"result"`
-	Error  Error              `json:"error"`
+// GetBlockResponse terra response from /block
+
+type GetBlockResponseV4 struct {
+	// ID     string        `json:"id"`
+	RPC    string        `json:"jsonrpc"`
+	Result ResultBlockV4 `json:"result"`
+	Error  Error         `json:"error"`
 }
 
-// ResultBlockchainV4 is result of fetching block
-type ResultBlockchainV4 struct {
-	LastHeight string        `json:"last_height"`
-	BlockMetas []BlockMetaV4 `json:"block_metas"`
+type ResultBlockV4 struct {
+	Block   BlockV4 `json:"block"`
+	BlockID BlockID `json:"block_id"`
 }
 
-// BlockMetaV4 is block metadata
-type BlockMetaV4 struct {
-	BlockID BlockID       `json:"block_id"`
-	Header  BlockHeaderV4 `json:"header"`
-	NumTxs  string        `json:"num_txs"`
+// BlockV4 is terra block data
+type BlockV4 struct {
+	Header BlockHeaderV4 `json:"header"`
+	Data   BlockDataV4   `json:"data"`
 }
 
-// BlockMetaV4 is block header
+type BlockDataV4 struct {
+	Txs []string `json:"txs"`
+}
+
+// BlockHeaderV4 is block header
 type BlockHeaderV4 struct {
 	Height  string `json:"height"`
 	ChainID string `json:"chain_id"`
 	Time    string `json:"time"`
+}
+
+// DelegationResponse is terra response for querying /delegations
+type DelegationResponseV4 struct {
+	Height      string         `json:"height"`
+	Delegations []DelegationV4 `json:"result"`
+}
+
+type DelegationV4 struct {
+	DelegatorAddress string  `json:"delegator_address"`
+	ValidatorAddress string  `json:"validator_address"`
+	Shares           string  `json:"shares"`
+	Balance          Balance `json:"balance"`
+}
+
+type Balance struct {
+	Denom  string `json:"denom"`
+	Amount string `json:"amount"`
 }